# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion Comptable",
        "description": "API pour la gestion comptable avec journal des écritures et génération de balance comptable",
        "contact": {
            "name": "Support API",
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Récupérer la liste de tous les comptes comptables",
                "description": "Afficher une liste des comptes.",
                "operationId": "7b230c67fcba4b29cb17cef21ff2240b",
                "responses": {
                    "200": {
                        "description": "Liste des comptes comptables",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compte"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Créer un nouveau compte comptable",
                "description": "Enregistrer un nouveau compte.",
                "operationId": "a343f058b23d5e86ac7bbb8729ea6be9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "code",
                                    "type"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Banque XYZ"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "512100"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "actif",
                                            "passif",
                                            "produit",
                                            "charge"
                                        ],
                                        "example": "actif"
                                    },
                                    "solde": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/comptes/{id}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Récupérer les détails d'un compte comptable",
                "description": "Afficher les détails d'un compte spécifique.",
                "operationId": "84b3305058d29a6e778d958bdbb20bbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Mettre à jour un compte comptable",
                "description": "Mettre à jour un compte spécifique.",
                "operationId": "719c17a9c7faeb938e5451323f3d15d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "code",
                                    "type"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Banque XYZ"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "512100"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "actif",
                                            "passif",
                                            "produit",
                                            "charge"
                                        ],
                                        "example": "actif"
                                    },
                                    "solde": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Supprimer un compte comptable",
                "description": "Supprimer un compte spécifique.",
                "operationId": "94f006a9ddc0f1c0f9b7c9fd56efafbf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Impossible de supprimer le compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Impossible de supprimer ce compte car il est utilisé dans des transactions"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Compte": {
                "title": "Compte",
                "description": "Modèle de compte comptable",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Banque"
                    },
                    "code": {
                        "type": "string",
                        "example": "512"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "actif",
                            "passif",
                            "produit",
                            "charge"
                        ],
                        "example": "actif"
                    },
                    "solde": {
                        "type": "number",
                        "format": "float",
                        "example": 1000
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}